Outsource spawn_handler_thread logic to HandlerMapping struct
